requirements.txt
pandas
numpy
openai
langchain
langchain_openai
pinecone-client
llama-parse
langchain-text-splitters